requirements.txt
deepseek-api>=1.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-mock>=3.6.1
requests>=2.26.0
markdown>=3.3.0
chardet>=4.0.0 